[CMYK-118] 코드 리펙토링
- sql로 table을 생성하는 파일이름 변경
- chatroom의 id값을 Long으로 변경

+++ src/main/java/com/cmyk/ego/speaktoyouspring/api/personalized_data/chatroom/ChatRoomDTO.java
@@ -15,7 +15,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ChatRoomDTO {
-    private Integer id; // 채팅방 ID
+    private Long id; // 채팅방 ID
 
     @NotBlank(message = "UID는 필수입니다.")
     private String uid; // 사용자 UID
